+++ vcl/inc/salframe.hxx
@@ -240,7 +240,7 @@ public:
     // that should not know more than necessary about the SalFrame implementation
     // (e.g. input methods, printer update handlers).
     long                    CallCallback( sal_uInt16 nEvent, const void* pEvent ) const
-        { return m_pProc ? m_pProc( m_pWindow, const_cast<SalFrame*>(this), nEvent, pEvent ) : 0; }
+        { return m_pProc ? long(m_pProc( m_pWindow, const_cast<SalFrame*>(this), nEvent, pEvent )) : 0; }
 };
 
 #endif // INCLUDED_VCL_INC_SALFRAME_HXX
